+++ geocoding_darwin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Updates the example app to demonstrate Darwin-specific native CLGeocoder
+  methods via `clgeocoder.dart`.
+
 ## 1.0.0
 
 - Initial release of the geocoding_darwin package containing easy geocoding and